+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/expressions/DifferentiableExpression.kt
@@ -5,9 +5,6 @@
 
 package space.kscience.kmath.expressions
 
-import space.kscience.kmath.misc.StringSymbol
-import space.kscience.kmath.misc.Symbol
-
 /**
  * Represents expression which structure can be differentiated.
  *